WebScrapping/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.14
demjson==2.2.4
feedfinder2==0.0.4
feedparser==6.0.1
filelock==3.0.12
gensim==3.8.3
idna==2.10
jieba3k==0.35.1
joblib==0.17.0
jtutils==0.0.8
kiwisolver==1.2.0
leven==1.0.4
lxml==4.6.1
matplotlib==3.3.2
newspaper3k==0.2.8
nltk==3.5
nose==1.3.7
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.1
pyparsing==2.4.7
python-csv==0.0.13
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.23
requests==2.24.0
requests-file==1.5.1
scipy==1.5.3
selenium==3.141.0
sgmllib3k==1.0.0
six==1.15.0
smart-open==3.0.0
soupsieve==2.0.1
tinysegmenter==0.3
tldextract==3.0.2
tqdm==4.50.2
urllib3==1.25.11
xlrd==1.2.0
XlsxWriter==1.3.7
xmltodict==0.12.0
